fix(snippetz): resolve axios migration regressions
